다민이사이트

+++ demo/src/main/java/com/example/MyWeb/demo/model/User.java
@@ -17,7 +17,7 @@ public class User {
 
     @Id //primary key
     @GeneratedValue(strategy = GenerationType.IDENTITY) // 프로젝트에서 연결된 DB의 넘버링 전략을 따른다.(AutoIncrement)
-    private int id; //시퀀스 auto_increment
+    private Long id; //시퀀스 auto_increment
 
     @Column(nullable = false,length = 30)
     private String username; //아이디
