+++ demo/src/main/java/com/example/MyWeb/demo/repository/BoardRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 //@Repository 생략가능 jparepository를 extends해서
-public interface BoardRepository extends JpaRepository<Board, Integer> { //user table의 primary key는 integer
+public interface BoardRepository extends JpaRepository<Board, Long> { //user table의 primary key는 integer
     //SELECT * FROM user WHERE username=1?;
 
 }
